added tests for gini indices and moved f1measure tests to paircounting test class
