+++ elki-clustering/src/test/java/elki/evaluation/clustering/PairCountingTest.java
@@ -24,10 +24,23 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
+import elki.algorithm.AbstractSimpleAlgorithmTest;
+import elki.clustering.trivial.ByLabelClustering;
+import elki.clustering.trivial.TrivialAllInOne;
+import elki.clustering.trivial.TrivialAllNoise;
+import elki.data.Clustering;
+import elki.data.model.Model;
+import elki.database.Database;
 import elki.database.ids.DBIDRange;
 import elki.database.ids.DBIDUtil;
 
 public class PairCountingTest {
+
+  // the following values depend on the data set used!
+  String dataset = "elki/testdata/unittests/hierarchical-3d2d1d.csv";
+
+  // size of the data set
+  int shoulds = 600;
 
   @Test
   public void testPairCountingSKlearn() {
@@ -44,7 +57,30 @@ public class PairCountingTest {
     assertEquals("F1-Measure not as expected", 0.487804878048781, pc.f1Measure(), 1e-15);
     assertEquals("Fb(5)-Measure not as expected", 0.499040307101727, pc.fMeasure(5), 1e-15);
     assertEquals("Adjusted Rand-Index not as expected", 0.26694045174538, pc.adjustedRandIndex(), 1e-15);
-    assertEquals("Mirkin-Index not as expected", 168, pc.mirkin());
+    assertEquals("Mirkin-Index not as expected", 168, pc.mirkin());// divide by
+                                                                   // 2
+  }
+
+  @Test
+  public void compareDatabases() {
+    Database db = AbstractSimpleAlgorithmTest.makeSimpleDatabase(dataset, shoulds);
+
+    Clustering<Model> rai = new TrivialAllInOne().autorun(db);
+    Clustering<Model> ran = new TrivialAllNoise().autorun(db);
+    Clustering<?> rbl = new ByLabelClustering().autorun(db);
+
+    assertEquals(1.0, computeFMeasure(rai, rai, false), Double.MIN_VALUE);
+    assertEquals(1.0, computeFMeasure(ran, ran, false), Double.MIN_VALUE);
+    assertEquals(1.0, computeFMeasure(rbl, rbl, false), Double.MIN_VALUE);
+
+    assertEquals(0.009950248756218905, computeFMeasure(ran, rbl, true), Double.MIN_VALUE);
+    assertEquals(0.0033277870216306157, computeFMeasure(rai, ran, true), Double.MIN_VALUE);
+
+    assertEquals(0.5 /* 0.3834296724470135 */, computeFMeasure(rai, rbl, false), Double.MIN_VALUE);
+  }
+
+  private double computeFMeasure(Clustering<?> c1, Clustering<?> c2, boolean noise) {
+    return new ClusterContingencyTable(true, noise, c1, c2).getPaircount().f1Measure();
   }
 
 }
